+++ src/test/java/step3/RaceCarTest.java
@@ -1,12 +1,8 @@
 package step3;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.operation.InputValue;
-import step3.operation.Operation;
 import step3.view.InputView;
 import step3.view.ResultView;
-
-import java.util.Random;
 
 public class RaceCarTest {
 
@@ -21,12 +17,12 @@ public class RaceCarTest {
     @DisplayName("최종 테스트")
     @Test
     public void 경주() {
-        InputView iv = new InputView();
-        int numOfCars = iv.numOfCars("5");
-        int numOfAtp = iv.numberOfAttempts("10");
+        InputView inputView = new InputView();
+        int numOfCars = inputView.numOfCars("3");
+        int numOfAtp = inputView.numberOfAttempts("5");
 
-        ResultView rv = new ResultView();
-        rv.resultView(numOfAtp, numOfCars);
+        ResultView resultView = new ResultView();
+        resultView.resultView(numOfAtp, numOfCars);
 
     }
